requirements.txt
# Buggy Call - Requirements
# Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5

# Database
PyMySQL==1.1.0
mysqlclient==2.2.0

# Validation & Serialization
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Security
bcrypt==4.1.1
cryptography==41.0.7
Flask-Limiter==3.5.0
Flask-WTF==1.2.1
Flask-Talisman==1.1.0
flask-swagger-ui==5.21.0

# QR Code & Image
qrcode==7.4.2
Pillow==10.1.0

# Push Notifications
pywebpush==1.14.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# WebSocket
python-socketio==5.10.0
eventlet==0.33.3

# Caching
redis==5.0.1
Flask-Caching==2.1.0

# Server
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0
faker==20.1.0

# Reporting & Export
openpyxl==3.1.2
reportlab==4.0.7
